Global terrorism data analysis using python
A terrorism dataset with various information like day of the attack, year of attack, region, country, state, city, etc., was successfully visualized using a dash and plotly as a major python framework. A web application consisting of a map and a chart tool was created based on python. Libraries like pandas,and numpy were used for dealing with the data frame. Visualization of the data makes it easy for the client to analyze the data and get the desired information.

The development of the web app initially aims towards creating 
Map tool
Chart tool

The map tool consists of 
World map tool
India map tool


The chart tool consists of :
World chart tool
India chart tool

Followed by there are 7 dropdowns created which are as follows 
Month 
Day
Region
Country
State
City
Type of attack

A year range slide is also created after the dropdowns for the years in the range of 1970-2018

Finally, a global terrorism map is created using a visualization map tool express by plotly.